cleaned up repo and moved files to correct places, also renamed files for clarity
