+++ environment.yml
@@ -3,3 +3,145 @@ channels:
   - conda-forge
   - defaults
 dependencies:
+  - _py-xgboost-mutex=2.0=cpu_0
+  - appnope=0.1.0=py36hf537a9a_0
+  - asn1crypto=0.24.0=py36_0
+  - backcall=0.1.0=py36_0
+  - blas=1.0=mkl
+  - bleach=3.1.0=py36_0
+  - boto=2.49.0=py36_0
+  - boto3=1.9.234=py_0
+  - botocore=1.12.234=py_0
+  - branca=0.3.1=py_0
+  - bz2file=0.98=py36_1
+  - ca-certificates=2019.11.27=0
+  - certifi=2019.11.28=py36_0
+  - cffi=1.11.5=py36h6174b99_1
+  - chardet=3.0.4=py36_1003
+  - cryptography=2.3.1=py36hdbc3d79_0
+  - cryptography-vectors=2.3.1=py36_0
+  - cycler=0.10.0=py36hfc81398_0
+  - cython=0.28.5=py36h0a44026_0
+  - decorator=4.3.0=py36_0
+  - defusedxml=0.6.0=py_0
+  - docutils=0.15.2=py36_0
+  - entrypoints=0.3=py36_0
+  - freetype=2.9.1=hb4e5f40_0
+  - gensim=3.4.0=py36h1de35cc_0
+  - h5py=2.9.0=py36h3134771_0
+  - hdf5=1.10.4=hfa1e0ec_0
+  - idna=2.7=py36_0
+  - intel-openmp=2018.0.3=0
+  - ipykernel=5.1.2=py36h39e3cac_0
+  - ipyleaflet=0.11.4=py36_1
+  - ipython=6.5.0=py36_0
+  - ipython_genutils=0.2.0=py36h241746c_0
+  - ipywidgets=7.5.1=py_0
+  - jedi=0.12.1=py36_0
+  - jinja2=2.10.3=py_0
+  - jmespath=0.9.4=py_0
+  - joblib=0.14.0=py_0
+  - jsonschema=2.6.0=py36hb385e00_0
+  - jupyter_client=5.3.3=py36_1
+  - jupyter_core=4.4.0=py36_0
+  - kiwisolver=1.1.0=py36h0a44026_0
+  - krb5=1.16.1=h24a3359_6
+  - libcxx=4.0.1=hcfea43d_1
+  - libcxxabi=4.0.1=hcfea43d_1
+  - libedit=3.1.20170329=hb402a30_2
+  - libffi=3.2.1=h475c297_4
+  - libgfortran=3.0.1=h93005f0_2
+  - libpng=1.6.37=ha441bb4_0
+  - libpq=10.5=hf30b1f0_0
+  - libsodium=1.0.16=h3efe00b_0
+  - libxgboost=0.72=h0a44026_0
+  - llvm-openmp=4.0.1=hcfea43d_1
+  - markupsafe=1.1.1=py36h1de35cc_0
+  - matplotlib=3.0.2=py36h54f8f79_0
+  - mistune=0.8.4=py36h1de35cc_0
+  - mkl=2019.4=233
+  - mkl-service=2.3.0=py36hfbe908c_0
+  - mkl_fft=1.0.14=py36h5e564d8_0
+  - mkl_random=1.1.0=py36ha771720_0
+  - nb_conda=2.2.1=py36_0
+  - nb_conda_kernels=2.2.2=py36_0
+  - nbconvert=5.6.0=py36_1
+  - nbformat=4.4.0=py36h827af21_0
+  - ncurses=6.1=h0a44026_1
+  - networkx=2.2=py36_1
+  - nltk=3.4.5=py36_0
+  - notebook=6.0.1=py36_0
+  - numpy=1.17.2=py36h99e6662_0
+  - numpy-base=1.17.2=py36h6575580_0
+  - openssl=1.0.2u=h1de35cc_0
+  - pandas=0.23.4=py36h6440ff4_0
+  - pandoc=2.2.3.2=0
+  - pandocfilters=1.4.2=py36_1
+  - parso=0.3.1=py36_0
+  - patsy=0.5.1=py36_0
+  - pexpect=4.6.0=py36_0
+  - pickleshare=0.7.4=py36hf512f8e_0
+  - pip=10.0.1=py36_0
+  - plotly=3.1.0=py36h28b3542_0
+  - prometheus_client=0.7.1=py_0
+  - prompt_toolkit=1.0.15=py36haeda067_0
+  - psycopg2=2.7.5=py36hdbc3d79_0
+  - ptyprocess=0.6.0=py36_0
+  - py-xgboost=0.72=py36h0a44026_0
+  - py4j=0.10.7=py36_0
+  - pycparser=2.18=py36_1
+  - pydotplus=2.0.2=py36_1
+  - pygments=2.2.0=py36h240cd3f_0
+  - pymongo=3.9.0=py36h0a44026_0
+  - pyopenssl=18.0.0=py36_0
+  - pyparsing=2.4.2=py_0
+  - pysocks=1.6.8=py36_0
+  - pyspark=2.4.0=py36_0
+  - python=3.6.6=hc167b69_0
+  - python-dateutil=2.8.0=py36_0
+  - pytz=2018.5=py36_0
+  - pyzmq=18.1.0=py36h0a44026_0
+  - readline=7.0=h1de35cc_5
+  - requests=2.19.1=py36_0
+  - retrying=1.3.3=py36_2
+  - s3transfer=0.2.1=py36_0
+  - scikit-learn=0.22=py36h27c97d8_0
+  - scipy=1.1.0=py36h1410ff5_2
+  - seaborn=0.9.0=py36_0
+  - send2trash=1.5.0=py36_0
+  - setuptools=40.0.0=py36_0
+  - simplegeneric=0.8.1=py36_2
+  - six=1.11.0=py36_1
+  - smart_open=1.8.4=py_0
+  - soupsieve=1.9.3=py36_0
+  - sqlalchemy=1.3.11=py36h1de35cc_0
+  - sqlite=3.24.0=ha441bb4_0
+  - statsmodels=0.9.0=py36h1d22016_0
+  - terminado=0.8.2=py36_0
+  - testpath=0.4.2=py36_0
+  - tk=8.6.7=h35a86e2_3
+  - tornado=6.0.3=py36h1de35cc_0
+  - tqdm=4.36.1=py_0
+  - traitlets=4.3.2=py36h65bd3ce_0
+  - traittypes=0.2.1=py_1
+  - urllib3=1.23=py36_0
+  - wcwidth=0.1.7=py36h8c6ec74_0
+  - webencodings=0.5.1=py36_1
+  - wheel=0.31.1=py36_0
+  - widgetsnbextension=3.5.1=py36_0
+  - xarray=0.13.0=py_0
+  - xlrd=1.1.0=py36_1
+  - xz=5.2.4=h1de35cc_4
+  - zeromq=4.3.1=h0a44026_3
+  - zlib=1.2.11=h1de35cc_3
+prefix: //anaconda3/envs/linreg-env
+​
+Collapse
+
+
+
+
+
+
+
+
